+++ test/hexapdf/encryption/test_standard_security_handler.rb
@@ -292,4 +292,30 @@ describe HexaPDF::Encryption::StandardSecurityHandler do
     @handler.set_up_encryption(permissions: perms)
     assert_equal([:copy_content, :modify_content], @handler.permissions.sort)
   end
+
+  describe "handling of metadata streams" do
+    before do
+      @doc = HexaPDF::Document.new
+      @doc.encrypt(encrypt_metadata: false)
+      @output = StringIO.new(''.b)
+    end
+
+    it "doesn't decrypt or encrypt the document level metadata stream if /EncryptMetadata is false" do
+      @doc.catalog[:Metadata] = @doc.wrap({Type: :Metadata, Subtype: :XML}, stream: "HELLODATA")
+      @doc.write(@output)
+      assert_match(/stream\nHELLODATA\nendstream/, @output.string)
+
+      doc = HexaPDF::Document.new(io: @output)
+      assert_equal('HELLODATA', doc.catalog[:Metadata].stream)
+    end
+
+    it "doesn't modify decryption/encryption for arbitrary metadata streams" do
+      @doc.catalog[:Anything] = @doc.wrap({Type: :Metadata, Subtype: :XML}, stream: "HELLODATA")
+      @doc.write(@output)
+      refute_match(/stream\nHELLODATA\nendstream/, @output.string)
+
+      doc = HexaPDF::Document.new(io: @output)
+      assert_equal('HELLODATA', doc.catalog[:Anything].stream)
+    end
+  end
 end